Move KotlinDetector location
