+++ core/src/main/java/feign/AsyncResponseHandler.java
@@ -18,7 +18,6 @@ import static feign.Util.ensureClosed;
 import feign.Logger.Level;
 import feign.codec.Decoder;
 import feign.codec.ErrorDecoder;
-import feign.kotlinSupport.KotlinDetector;
 
 import java.io.IOException;
 import java.lang.reflect.Type;
