+++ core/src/main/java/feign/KotlinDetector.java
@@ -1,4 +1,4 @@
-package feign.kotlinSupport;
+package feign;
 
 import kotlin.Unit;
